# Use constant for baggage validation bitsets