--- a/api/all/src/main/java/io/opentelemetry/api/baggage/propagation/Element.java
+++ b/api/all/src/main/java/io/opentelemetry/api/baggage/propagation/Element.java
@@ -14,6 +14,21 @@ import java.util.BitSet;
  */
 class Element {
 
+  private static final BitSet EXCLUDED_KEY_CHARS = new BitSet(128);
+  private static final BitSet EXCLUDED_VALUE_CHARS = new BitSet(128);
+
+  static {
+    for (char c :
+        new char[] {
+          '(', ')', '<', '>', '@', ',', ';', ':', '\\', '"', '/', '[', ']', '?', '=', '{', '}'
+        }) {
+      EXCLUDED_KEY_CHARS.set(c);
+    }
+    for (char c : new char[] {'"', ',', ';', '\\'}) {
+      EXCLUDED_VALUE_CHARS.set(c);
+    }
+  }
+
   private final BitSet excluded;
 
   private boolean leadingSpace;
@@ -22,6 +37,14 @@ class Element {
   private int start;
   private int end;
   private String value;
+
+  static Element createKeyElement() {
+    return new Element(EXCLUDED_KEY_CHARS);
+  }
+
+  static Element createValueElement() {
+    return new Element(EXCLUDED_VALUE_CHARS);
+  }
 
   /**
    * Constructs element instance.