Randomly noticed. Just small improvement

After
```
Benchmark                                                                Mode  Cnt      Score      Error   Units
W3CBaggagePropagatorBenchmark.largeBaggage                               avgt   45     42.915 ±    0.496   us/op
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.alloc.rate                avgt   45    368.208 ±    4.159  MB/sec
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.alloc.rate.norm           avgt   45  24872.049 ±    0.062    B/op
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.churn.G1_Eden_Space       avgt   45    370.006 ±   25.219  MB/sec
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.churn.G1_Eden_Space.norm  avgt   45  24992.676 ± 1681.603    B/op
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.churn.G1_Old_Gen          avgt   45      0.010 ±    0.002  MB/sec
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.churn.G1_Old_Gen.norm     avgt   45      0.706 ±    0.146    B/op
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.count                     avgt   45    165.000             counts
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.time                      avgt   45     67.000                 ms
W3CBaggagePropagatorBenchmark.smallBaggage                               avgt   45      1.923 ±    0.053   us/op
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.alloc.rate                avgt   45    452.804 ±   12.043  MB/sec
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.alloc.rate.norm           avgt   45   1368.002 ±    0.003    B/op
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.churn.G1_Eden_Space       avgt   45    450.834 ±   26.759  MB/sec
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.churn.G1_Eden_Space.norm  avgt   45   1361.720 ±   69.804    B/op
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.churn.G1_Old_Gen          avgt   45      0.001 ±    0.001  MB/sec
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.churn.G1_Old_Gen.norm     avgt   45      0.003 ±    0.002    B/op
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.count                     avgt   45    201.000             counts
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.time                      avgt   45     77.000                 ms
```

Before
```
Benchmark                                                                Mode  Cnt      Score      Error   Units
W3CBaggagePropagatorBenchmark.largeBaggage                               avgt   45     44.097 ±    0.276   us/op
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.alloc.rate                avgt   45    360.876 ±    2.217  MB/sec
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.alloc.rate.norm           avgt   45  25064.050 ±    0.063    B/op
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.churn.G1_Eden_Space       avgt   45    363.243 ±   26.208  MB/sec
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.churn.G1_Eden_Space.norm  avgt   45  25228.640 ± 1814.921    B/op
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.churn.G1_Old_Gen          avgt   45      0.005 ±    0.001  MB/sec
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.churn.G1_Old_Gen.norm     avgt   45      0.376 ±    0.097    B/op
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.count                     avgt   45    162.000             counts
W3CBaggagePropagatorBenchmark.largeBaggage:·gc.time                      avgt   45     65.000                 ms
W3CBaggagePropagatorBenchmark.smallBaggage                               avgt   45      2.043 ±    0.049   us/op
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.alloc.rate                avgt   45    485.899 ±   12.099  MB/sec
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.alloc.rate.norm           avgt   45   1560.002 ±    0.003    B/op
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.churn.G1_Eden_Space       avgt   45    486.558 ±   26.021  MB/sec
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.churn.G1_Eden_Space.norm  avgt   45   1561.931 ±   74.079    B/op
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.churn.G1_Old_Gen          avgt   45      0.001 ±    0.001  MB/sec
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.churn.G1_Old_Gen.norm     avgt   45      0.004 ±    0.002    B/op
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.count                     avgt   45    217.000             counts
W3CBaggagePropagatorBenchmark.smallBaggage:·gc.time                      avgt   45     84.000                 ms
```